renamed main test files
